+++ EventSubscriber/AggregateSubscriber.php
@@ -7,9 +7,11 @@ namespace RevisionTen\CMS\EventSubscriber;
 use RevisionTen\CMS\Model\File;
 use RevisionTen\CMS\Model\Menu;
 use RevisionTen\CMS\Model\Role;
+use RevisionTen\CMS\Model\UserAggregate;
 use RevisionTen\CMS\Services\FileService;
 use RevisionTen\CMS\Services\MenuService;
 use RevisionTen\CMS\Services\RoleService;
+use RevisionTen\CMS\Services\UserService;
 use RevisionTen\CQRS\Event\AggregateUpdatedEvent;
 use Symfony\Component\EventDispatcher\EventSubscriberInterface;
 
@@ -24,18 +26,23 @@ class AggregateSubscriber implements EventSubscriberInterface
     /** @var FileService $fileService */
     private $fileService;
 
+    /** @var UserService $userService */
+    private $userService;
+
     /**
      * AggregateSubscriber constructor.
      *
      * @param MenuService $menuService
      * @param RoleService $roleService
      * @param FileService $fileService
+     * @param UserService $userService
      */
-    public function __construct(MenuService $menuService, RoleService $roleService, FileService $fileService)
+    public function __construct(MenuService $menuService, RoleService $roleService, FileService $fileService, UserService $userService)
     {
         $this->menuService = $menuService;
         $this->roleService = $roleService;
         $this->fileService = $fileService;
+        $this->userService = $userService;
     }
 
     public static function getSubscribedEvents()
@@ -50,16 +57,17 @@ class AggregateSubscriber implements EventSubscriberInterface
         /** @var \RevisionTen\CQRS\Interfaces\EventInterface $event */
         $event = $aggregateUpdatedEvent->getEvent();
 
-        if ($event->getCommand()->getAggregateClass() === Menu::class) {
-            $this->menuService->updateMenuRead($event->getCommand()->getAggregateUuid());
-        }
+        $aggregateClass = $event->getCommand()->getAggregateClass();
+        $aggregateUuid = $event->getCommand()->getAggregateUuid();
 
-        if ($event->getCommand()->getAggregateClass() === Role::class) {
-            $this->roleService->updateRoleRead($event->getCommand()->getAggregateUuid());
-        }
-
-        if ($event->getCommand()->getAggregateClass() === File::class) {
+        if ($aggregateClass === Menu::class) {
+            $this->menuService->updateMenuRead($aggregateUuid);
+        } elseif ($aggregateClass === Role::class) {
+            $this->roleService->updateRoleRead($aggregateUuid);
+        } elseif ($aggregateClass === File::class) {
             $this->fileService->updateFileRead($event->getCommand()->getAggregateUuid());
+        } elseif ($aggregateClass === UserAggregate::class) {
+            $this->userService->updateUserRead($event->getCommand()->getAggregateUuid());
         }
     }
 }